Initial design for unified memcard storage
Keeps PS1 + PS2 memcards together in one array, since both need to be accessible for PS2 operation

+++ pcsx2/Memcard/PS2/MemcardPS2.cpp
@@ -8,7 +8,7 @@
 #include <string>
 #include "DirectoryHelper.h"
 
-MemcardPS2::MemcardPS2(int port, int slot)
+MemcardPS2::MemcardPS2(size_t port, size_t slot)
 {
 	this->port = port;
 	this->slot = slot;
